Applied cheap, terrible hack to prevent escaped HTML symbols getting reverted before sending.
For some reason, after correct escaping of symbols like "&" and "á" by the htmlentities.py script,
the escaped symbols ("&amp;" and "&aacute;" respectively) symbols were getting back-converted to
raw symbols before getting sent to the XML-RPC script on the wordpress end of things. This lead
to an obscure "parse error. not well formed" bug from the server.

To solve this in the short term, I've just applied double-escaping in the python script, so that
an ampersand symbol "&" becomes "&amp;amp;", which gets converted by whatever's doing the back-
conversions to "&amp;" right before embedding in XML and sending to the server.